shape-reconstruction/implicit-differentiable-renderer/requirements.txt
ruamel.yaml
pybind11
python-config
tqdm
matplotlib
opencv-python
imageio
scikit-image
open3d
trimesh

